refactored controllers to service and removed controller dir
